# api-doc/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI Image Generator",
    "description": "",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/MrTimeey/ai-image-generator"
    },
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://github.com/MrTimeey/ai-image-generator/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
    "/api/openai/generate-images": {
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Generates image",
        "description": "Generates the image based on provided input parameters",
        "requestBody": {
          "description": "The generation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Information of the generated image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedImage"
                }
              }
            }
          }
        }
      }
    },
    "/api/openai/generate-alternative-images": {
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Generates alternative image",
        "description": "Generates alternatives based on provided image",
        "requestBody": {
          "description": "The alternative generation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAlternativesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Information of the generated image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedImage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GenerateImageRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "The description of the image",
            "example": "A friendly-looking dinosaur wearing a hoodie while running on a treadmill, digital art",
            "required": true
          },
          "amount": {
            "type": "integer",
            "description": "The amount of image variations ",
            "example": 1,
            "minimum": 1,
            "maximum": 10,
            "default": 1,
            "required": false
          },
          "size": {
            "type": "string",
            "enum": [ "SMALL", "MEDIUM", "LARGE"],
            "description": "The size of the generated image. Possible options: 'SMALL' (256x256), 'MEDIUM' (512x512), 'LARGE' (1024x1024)",
            "required": false,
            "default": "SMALL",
            "example": "SMALL"
          }
        }
      },
      "GenerateAlternativesRequest": {
        "type": "object",
        "properties": {
          "baseImage": {
            "type": "string",
            "description": "The content of an image tag",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAwAAAABCAIAAABlidhuAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAJcEhZcwAAEnQAABJ0Ad5mH3gAAAARSURBVBhXY/z//z8DfsDAAABo4QL/OfELpAAAAABJRU5ErkJggg==",
            "required": true
          },
          "amount": {
            "type": "integer",
            "description": "The amount of image variations ",
            "example": 1,
            "minimum": 1,
            "maximum": 10,
            "default": 1,
            "required": false
          },
          "size": {
            "type": "string",
            "enum": [ "SMALL", "MEDIUM", "LARGE"],
            "description": "The size of the generated image. Possible options: 'SMALL' (256x256), 'MEDIUM' (512x512), 'LARGE' (1024x1024)",
            "required": false,
            "default": "SMALL",
            "example": "SMALL"
          }
        }
      },
      "GeneratedImage": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "description": "The timestamp of the image creation",
            "example": "15-12-2022_11-57-34"
          },
          "urls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageUrl"
            }
          }
        }
      },
      "ImageUrl": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID",
            "example": "31b353eb-74c7-4f16-b411-5a3367afb518"
          },
          "url": {
            "type": "string",
            "description": "The url to the generated image"
          }
        }
      }
    }
  }
}